Working on docs.
Bug: a linfree() A with fills_slab() == 1 could free s while a linfree() B
which cleared st.lost is still pushing s to her->slabs.

There's probably no choice but for B to prevent fills_slab() == 1 until
the push completes. Something like "don't push if a failed clear_cas_won
shows fills_slab(st)" still fails if s is freed and reallocated. A version
count might let B identify that case, but it's not going to be worth it.

It suffices to not push b onto s->hot_blocks if clearing st.lost. Push it
onto s->local_blocks directly, and omit the
lfstack_clear_cas_won()-lfstack_convert() dance. Bad thing is that
linalloc() may have to clear s->hot_blocks a little sooner (because you
didn't). But this is unlikely because s->hot_blocks was empty when you
cleared st.lost. So the good thing is that this pointless and expensive
clear_cas_won disappears.